Disable actions for users that have no permission

+++ admin/company-index.php
@@ -23,26 +23,31 @@
         <div class="container-fluid">
             <div class="row">
                 <div class="col-md-12">
-                    <div class="page-header clearfix">
-                        <h2 class="float-left">Companies Details</h2>
-                        <a href="company-create.php" class="btn btn-success float-right">Add New Record</a>
-                        <a href="company-index.php" class="btn btn-info float-right mr-2">Reset View</a>
-                        <a href="index.php" class="btn btn-secondary float-right mr-2">Back</a>
-                    </div>
-
-                    <div class="form-row">
-                        <form action="company-index.php" method="get">
-                        <div class="col">
-                          <input type="text" class="form-control" placeholder="Search this table" name="search">
-                        </div>
-                    </div>
-                        </form>
-                    <br>
-
                     <?php
                     // Include config file
                     require_once "config.php";
                     require_once "helpers.php";
+                    require_once('permissions.php');
+
+                    echo "<div class='page-header clearfix'>";
+                        echo "<h2 class='float-left'>Companies Details</h2>";
+                        if ($_SESSION['SFx3'] == 1) {
+                            echo "<a href='company-create.php' class='btn btn-success float-right'>Add New Company</a>";
+                        }
+                        echo "<a href='company-index.php' class='btn btn-info float-right mr-2'>Reset View</a>";
+                        echo "<a href='index.php' class='btn btn-secondary float-right mr-2'>Back</a>";
+                    echo "</div>";
+                    
+                    echo "<div class='form-row'>";
+                        echo "<form action='company-index.php' method='get'>";
+                            echo "<div class='col'>";
+                            if ($_SESSION['SFx2'] == 1) {
+                                echo "<input type='text' class='form-control' placeholder='Search this table' name='search'>";
+                            } 
+                            echo "</div>";
+                        echo "</form>";
+                    echo "</div>";
+                    echo "<br>"; 
 
                     //Get current URL and parameters for correct pagination
                     $protocol = $_SERVER['SERVER_PROTOCOL'];
@@ -135,8 +140,12 @@
                                     echo "<td>" . $row['id'] . "</td>";echo "<td>" . $row['name'] . "</td>";echo "<td>" . $row['email'] . "</td>";echo "<td>" . $row['sector'] . "</td>";
                                         echo "<td>";
                                             echo "<a href='company-read.php?id=". $row['id'] ."' title='View Record' data-toggle='tooltip'><i class='far fa-eye'></i></a>";
-                                            echo "<a href='company-update.php?id=". $row['id'] ."' title='Update Record' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
-                                            echo "<a href='company-delete.php?id=". $row['id'] ."' title='Delete Record' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+                                            if ($_SESSION['SFx4'] == 1) {
+                                                echo "<a href='company-update.php?id=". $row['id'] ."' title='Update Record' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
+                                            }
+                                            if ($_SESSION['SFx6'] == 1) {
+                                                echo "<a href='company-delete.php?id=". $row['id'] ."' title='Delete Record' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+                                            }
                                         echo "</td>";
                                     echo "</tr>";
                                 }
